minor: isolation of featureItem
